feat: move o middleware do inertia pra uma função utilitária & injeta os temas (do Projetário e do sistema do usuário) na view data & nas propriedades das páginas
