+++ src/server.rs
@@ -1,28 +1,19 @@
-use std::sync::Arc;
-
-use actix_session::{SessionExt, SessionMiddleware};
+use actix_session::SessionMiddleware;
+use actix_web::App;
 use actix_web::body::{BoxBody, EitherBody};
 use actix_web::cookie::{Key, SameSite};
 use actix_web::dev::{ServiceFactory, ServiceRequest, ServiceResponse};
 use actix_web::middleware::NormalizePath;
-use actix_web::{App, HttpMessage};
 use config::app::{AppConfig, RustEnv};
-use dominio::identidade::traits::IntoUsuarioModelo;
-use futures_util::FutureExt;
-use inertia_rust::actix::InertiaMiddleware;
-use inertia_rust::{InertiaProp, InertiaService, IntoInertiaPropResult, hashmap};
+use inertia_rust::InertiaService;
 use inertia_sessions::file_session::FileSessionStore;
 use inertia_sessions::middlewares::garbage_collector::GarbageCollectorMiddleware;
 use inertia_sessions::middlewares::reflash_temporary_session::ReflashTemporarySessionMiddleware;
-use serde_json::Map;
 
 use crate::infra::http::RouterRegistrable;
-use crate::infra::http::middlewares::usuario_da_requisicao::{
-    MiddlewareUsuarioDaRequisicao,
-    UsuarioDaRequisicao,
-};
-use crate::infra::http::presenters::usuario_modelo::UsuarioModeloPresenter;
+use crate::infra::http::middlewares::usuario_da_requisicao::MiddlewareUsuarioDaRequisicao;
 use crate::infra::http::routers::web::WebRouter;
+use crate::libs::inertia::middleware::get_inertia_middleware;
 
 pub fn get_server() -> App<
     impl ServiceFactory<
@@ -39,33 +30,7 @@ pub fn get_server() -> App<
 
     App::new()
         .wrap(GarbageCollectorMiddleware)
-        .wrap(InertiaMiddleware::new().with_shared_props(Arc::new(|req| {
-            let usuario = req.extensions().get::<UsuarioDaRequisicao>().unwrap_or_else(|| {
-                log::warn!("Usuário da requisição não encontrada pelo `InertiaMiddleware`, caindo para o usuário convidado.");
-                &UsuarioDaRequisicao::Convidado
-            }).clone();
-
-            let usuario = match usuario {
-                UsuarioDaRequisicao::Convidado => None,
-                UsuarioDaRequisicao::Aluno(aluno) => Some(aluno.into_usuario_modelo()),
-                UsuarioDaRequisicao::Professor(professor) => Some(professor.into_usuario_modelo())
-            };
-
-            let autenticacao = usuario.map(|usuario| {
-                hashmap!["usuario".to_string() => UsuarioModeloPresenter::apresente(&usuario)]
-            });
-
-            let flash = req
-                .get_session()
-                .remove(app_config.sessions_flash_key)
-                .map(|flash_map| serde_json::from_str::<Map<_, _>>(&flash_map).unwrap_or_default())
-                .unwrap_or_default();
-
-            async { hashmap![
-                "flash" => InertiaProp::always(flash),
-                "autenticacao" => InertiaProp::Data(autenticacao.into_inertia_value()) 
-            ] }.boxed_local()
-        })))
+        .wrap(get_inertia_middleware())
         .wrap(ReflashTemporarySessionMiddleware)
         .wrap(MiddlewareUsuarioDaRequisicao)
         .wrap(
